add interface ILogic. use Open/Closed
