+++ Module_5/ILogicGame.cs
@@ -1,0 +1,13 @@
+﻿namespace Module_5
+{
+    interface ILogicGame
+    {
+        bool Status { get; set; }
+
+        string Message { get; set; }
+
+        void LogicGame();
+
+        void UpdateLevel();
+    }
+}
