+++ Module_5/IMaps.cs
@@ -1,7 +1,11 @@
-﻿namespace Module_5
+﻿using System.Collections.Generic;
+
+namespace Module_5
 {
     interface IMap
     {
+        List<ITrap> Traps { get;  }
+
         void CreateMap();
 
         void UpadateTrapOnMap();
