+++ Module_5/Logic.cs
@@ -1,110 +1,26 @@
-﻿using System.Collections.Generic;
+﻿using System;
 
 namespace Module_5
 {
     class Logic
-    {    
-        private readonly IPlayer player;
-        private readonly IPlayer quin;
-        private readonly IMap map;
-        private readonly List<ITrap> trap;
-        private readonly CommandByUser commandByUser;
-
-        public Logic(IPlayer player, IPlayer quin, IMap map, CommandByUser commandByUser, List<ITrap> trap)
+    {
+        public void LogicMoveGame(ILogicGame logic)
         {
-            this.player = player;
-            this.quin = quin;
-            this.map = map;
-            this.commandByUser = commandByUser;
-            this.trap = trap;
+            logic.LogicGame();        
         }
 
-        public bool Status { get; set; } = true;
-        public string Message { get; private set; }
-
-        public void LogicGame()
+        public void LogicUpdateLevelGame(ILogicGame logic)
         {
-            Move(commandByUser.DirectionPlayer);
-            OutOfBounds();
-            ContactWithTrap();
-            StatusGame();
-            ShowMessageAfterGameOver();
-        }
-
-        private void Move(CommandByUser.Direction direction)
-        {  
-            switch (direction)
+            if (!logic.Status)
             {
-                case CommandByUser.Direction.Left:
-                    player.PlayerPositionX -= 1;
-                    break;
-                case CommandByUser.Direction.Right:
-                    player.PlayerPositionX += 1;
-                    break;
-                case CommandByUser.Direction.Up:
-                    player.PlayerPositionY -= 1;
-                    break;
-                case CommandByUser.Direction.Down:
-                    player.PlayerPositionY += 1;
-                    break;
-            }
-        }       
-
-        private void OutOfBounds() 
-        {
-            if (player.PlayerPositionX >= map.Width)
-            {
-                player.PlayerPositionX = map.Width - 1;
-            }
-            else if (player.PlayerPositionX <= 0)
-            {
-                player.PlayerPositionX = 1;
-            }
-            else if (player.PlayerPositionY >= map.Height)
-            {
-                player.PlayerPositionY -=1;
-            }
-            else if (player.PlayerPositionY <= 0)
-            {
-                player.PlayerPositionY = 1;
-            }
-        }
-
-        private void ContactWithTrap()
-        {
-            foreach (var item in trap)
-            {
-                if ((player.PlayerPositionX == item.TrapPositionX) && (player.PlayerPositionY == item.TrapPositionY) && (item.TrapIsActive))
+                Console.WriteLine($"{logic.Message}\n" +
+                                   "Do you want play again?\n" +
+                                   "If yes, press any key. " +
+                                   "Else press Esc.");
+                if (Console.ReadKey().Key != ConsoleKey.Escape)
                 {
-                    player.PlayerHitPoints -=  item.TrapDamage;
-                    item.TrapIsActive = false;
-                    item.TrapIsVisible = true;
-                    break;
+                    logic.UpdateLevel();
                 }
-            }
-        }    
-        
-        private void StatusGame()
-        {
-            if (player.PlayerHitPoints <= 0)
-            {
-                Status = false;
-            }
-            if ((player.PlayerPositionX == quin.PlayerPositionX) && (player.PlayerPositionY == quin.PlayerPositionY))
-            {
-                Status = false;
-            }
-        }
-
-        private void ShowMessageAfterGameOver() 
-        {
-            if (player.PlayerHitPoints <= 0)
-            {
-                Message = "Game over. You lose.";
-            }
-            if ((player.PlayerPositionX == quin.PlayerPositionX) && (player.PlayerPositionY == quin.PlayerPositionY))
-            {
-                Message = "Game over. You win.";
             }
         }
     }
